change  type long to Long in some entites

+++ src/main/java/com/library/controller/CreanceController.java
@@ -52,12 +52,12 @@ public class CreanceController {
 
     @PostMapping(path = "/updateStatus")
     public ResponseEntity<Boolean> updateStatus(@RequestBody ObjectNode json) {
-        int reference;
+        Long reference;
         String ref;
         String status;
         try {
             ref = new ObjectMapper().treeToValue(json.get("ref"), String.class);
-            reference = Integer.parseInt(ref);
+            reference = Long.parseLong(ref);
             status = new ObjectMapper().treeToValue(json.get("status"), String.class);
             boolean test = this.creanceService.updateStatus(reference, status);
             if (test)
